finished working on uploading files

+++ controllers/admin.js
@@ -3,6 +3,8 @@ const Product = require("../models/product");
 const User = require("../models/user");
 const { validationResult } = require("express-validator/check");
 const errorController = require("../controllers/error");
+const fileHandler = require("../utils/file");
+const product = require("../models/product");
 
 //GET
 exports.getAddProduct = (req, res, next) => {
@@ -154,6 +156,7 @@ exports.postEditProduct = async (req, res, next) => {
     }
     product.title = updatedTitle;
     if (image) {
+        fileHandler.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
     }
     product.description = updatedDescription;
@@ -170,6 +173,11 @@ exports.postEditProduct = async (req, res, next) => {
 exports.postDeleteProduct = async (req, res, next) => {
     try {
         const productId = req?.body?.productId;
+        const product = await Product.findById(productId);
+        if (!product) {
+            return next(new Error("No product found"));
+        }
+        fileHandler.deleteFile(product.imageUrl);
         Product.deleteOne({ _id: productId }, (err, product) => {
             if (err) {
                 errorController.throwError(err, next);
@@ -180,6 +188,6 @@ exports.postDeleteProduct = async (req, res, next) => {
         });
         return res.redirect("/admin/products");
     } catch (error) {
-        errorController.throwError(error, next);
+        next(error);
     }
 };
